+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/models/ops.js
@@ -224,7 +224,7 @@ OpsPublishedDataCrawler = Marionette.Controller.extend({
         this.constituents = options.constituents;
     },
 
-    crawl: function() {
+    start: function() {
         var deferred = $.Deferred();
         var url_tpl = _.template('/api/ops/published-data/crawl/<%= constituents %>?query=<%= query %>');
         var url = url_tpl({constituents: this.constituents, query: this.query});
@@ -242,7 +242,7 @@ OpsPublishedDataCrawler = Marionette.Controller.extend({
                     deferred.reject('Empty response');
                 }
             }).error(function(error) {
-                deferred.reject('API failed', error);
+                deferred.reject('API failed: ' + JSON.stringify(error));
             });
         return deferred;
     },